Rename TestRetryProcessor class filename
